requirements.txt
flask
gevent
azure
gpiozero
AWSIoTPythonSDK
mysqlclient
picamera
twilio
mysql-connector